Refactor: 키오스크 추가 모달 분리
키오스크 추가 모달 분리

Ref: #5
